+++ RELEASE_NOTES.md
@@ -1,4 +1,9 @@
+#### 3.4.13
+* BUGFIX: Specialcase Gu.Reactive extension methods.
+
+#### 3.4.12
 * BUGFIX: When leaveOpen has default value
+
 #### 3.4.11
 * BUGFIX IDISP023 handle trivial and.
 * BUGFIX IDISP023 when chained constructors